requirements.txt
altgraph==0.17.5
APScheduler==3.6.3
cachetools==4.2.2
certifi==2026.2.25
joblib==1.5.3
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pefile==2024.8.26
psutil==7.2.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==13.15
pytz==2025.2
pywin32-ctypes==0.2.3
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tornado==6.1
tzdata==2025.3
tzlocal==5.3.1
